+++ collection_tests/stack_test.c
@@ -98,11 +98,11 @@ int main(void) {
             Stack stack = stack_new("1");
             test_empty_list(stack, stack);
             stack_push(stack, NEW_NODE(stack, 1));
-            obs_test(stack_length(stack), ==, (size_t)1);
+            obs_test(stack_length(stack), ==, (int)1);
             obs_test_false(stack_is_empty(stack));
 
             stack_free_node(stack_pop(stack));
-            obs_test(stack_length(stack), ==, (size_t)0);
+            obs_test(stack_length(stack), ==, (int)0);
             obs_test_true(stack_is_empty(stack));
             stack_free(stack);
         })
@@ -112,14 +112,14 @@ int main(void) {
             test_empty_list(stack, stack);
             for (int i = 0; i < 50; i++) {
                 stack_push(stack, NEW_NODE(stack, i));
-                obs_test(stack_length(stack), ==, (size_t)(i + 1));
+                obs_test(stack_length(stack), ==, (int)(i + 1));
                 obs_test_false(stack_is_empty(stack));
             }
 
             for (int i = 49; i >= 0; i--) {
                 StackNode n = stack_pop(stack);
                 obs_test(n->data.int_data, ==, (long long)i);
-                obs_test(stack_length(stack), ==, (size_t)i);
+                obs_test(stack_length(stack), ==, (int)i);
             }
             obs_test_true(stack_is_empty(stack));
         })
@@ -139,7 +139,7 @@ int main(void) {
             for (int i = 0; i < 100; i++) {
                 stack_push(stack, NEW_NODE(stack, i));
             }
-            obs_test(stack_length(stack), ==, (size_t)100);
+            obs_test(stack_length(stack), ==, (int)100);
             stack_clear(stack);
             test_empty_list(stack, stack);
         })
